+++ radical_translations/events/management/commands/export_events.py
@@ -1,0 +1,26 @@
+import csv
+
+from django.core.management.base import BaseCommand
+
+from radical_translations.events.models import Event
+
+
+class Command(BaseCommand):
+    help = "Exports `Event` data into a CSV file."
+
+    def handle(self, *args, **options):
+        self.stdout.write("Exporting Events into CSV file events.csv ...", ending=" ")
+        events = [r.to_dict() for r in Event.objects.all()]
+
+        if not events:
+            self.stderr.write(self.style.NOTICE("No events found!"))
+            return -1
+
+        fieldnames = events[0].keys()
+
+        with open("events.csv", "w") as f:
+            c = csv.DictWriter(f, fieldnames)
+            c.writeheader()
+            c.writerows(events)
+
+        self.stdout.write(self.style.SUCCESS("done"))