+++ radical_translations/events/models.py
@@ -1,5 +1,6 @@
 from typing import Dict, Optional
 
+from django.conf import settings
 from django.db import models
 from model_utils.models import TimeStampedModel
 
@@ -8,8 +9,10 @@ from radical_translations.core.models import Resource
 from radical_translations.utils.models import (
     Date,
     EditorialClassificationModel,
+    date_to_dict,
     get_geonames_place_from_gsx_place,
     get_gsx_entry_value,
+    place_to_dict_value,
 )
 
 # These models are based on the BIBFRAME 2.0 Event model
@@ -53,6 +56,16 @@ class Event(TimeStampedModel, EditorialClassificationModel):
         return "; ".join([c.label for c in self.classification.all()])
 
     get_classification.short_description = "Classification"  # type: ignore
+
+    def to_dict(self) -> Dict:
+        return {
+            "id": self.id,
+            **date_to_dict(self.date),
+            "place": place_to_dict_value(self.place),
+            "related_to": f"{settings.EXPORT_MULTIVALUE_SEPARATOR} ".join(
+                [r.to_dict_value() for r in self.related_to.all()]
+            ),
+        }
 
     @staticmethod
     def from_gsx_entry(entry: Dict[str, Dict[str, str]]) -> Optional["Event"]: